+++ packages/browser-wallet/src/shared/storage/access.ts
@@ -39,3 +39,4 @@ export const storedJsonRpcUrl = makeStorageAccessor<string>('local', ChromeStora
 export const storedSelectedAccount = makeStorageAccessor<string>('local', ChromeStorageKey.SelectedAccount);
 export const storedEncryptedSeedPhrase = makeStorageAccessor<EncryptedData>('local', ChromeStorageKey.SeedPhrase);
 export const storedTheme = makeStorageAccessor<Theme>('local', ChromeStorageKey.Theme);
+export const sessionPasscode = makeStorageAccessor<string>('session', ChromeStorageKey.Passcode);